Scrapped Everything and Started Over
This new method relies on the length of the coefficients list, so I removed the leading and trailing 0s.

I simplified my base functions for FFD, CFD, BFD, and RE so they would work with a function. finite_difference now mostly works with data. It knocks the first derivative out of the park and chooses which method to use, but when I pass that resultant data back through It gives me the right answer but divided by 1000.